broke up the tests into multiple files
